Moved the member model.
